Move file opening settings into own section
