+++ Packages/CmuxSettingsUI/Sources/CmuxSettingsUI/Scene/SettingsWindowScene.swift
@@ -427,7 +427,7 @@ public struct SettingsWindowRoot: View {
     @ViewBuilder
     private var sectionStack: some View {
         // Order matches the legacy in-app SettingsView scroll order:
-        // Account, App, Terminal, TextBox, Sidebar, Beta Features,
+        // Account, App, File Opening, Terminal, TextBox, Sidebar, Beta Features,
         // Automation, Browser (with embedded Import), Global Hotkey,
         // Keyboard Shortcuts, Workspace Colors, cmux.json, Reset.
         AccountSection(
@@ -443,6 +443,9 @@ public struct SettingsWindowRoot: View {
             hostActions: hostActions
         )
         .id(anchorID(for: .app))
+
+        FileOpeningSection(defaultsStore: defaultsStore, catalog: catalog)
+            .id(anchorID(for: .fileOpening))
 
         TerminalSection(
             defaultsStore: defaultsStore,
